refactoring of migration command
